Separate execution related classes
