+++ nautilus_trader/__init__.py
@@ -26,7 +26,7 @@ __author__ = "Nautech Systems"
 _MAJOR_VERSION = 1
 _MINOR_VERSION = 59
 _PATCH_VERSION = 0
-_PRE_RELEASE = ''
+_PRE_RELEASE = 'b0'
 
 __version__ = '.'.join([
     str(_MAJOR_VERSION),
