+++ nautilus_trader/common/execution_engine.pxd
@@ -1,0 +1,78 @@
+# -------------------------------------------------------------------------------------------------
+#  Copyright (C) 2015-2020 Nautech Systems Pty Ltd. All rights reserved.
+#  https://nautechsystems.io
+#
+#  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
+#  You may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+# -------------------------------------------------------------------------------------------------
+
+from nautilus_trader.common.account cimport Account
+from nautilus_trader.common.clock cimport Clock
+from nautilus_trader.common.execution_client cimport ExecutionClient
+from nautilus_trader.common.execution_database cimport ExecutionDatabase
+from nautilus_trader.common.logging cimport LoggerAdapter
+from nautilus_trader.common.portfolio cimport Portfolio
+from nautilus_trader.common.uuid cimport UUIDFactory
+from nautilus_trader.model.commands cimport Command
+from nautilus_trader.model.events cimport AccountStateEvent
+from nautilus_trader.model.events cimport Event
+from nautilus_trader.model.events cimport OrderCancelReject
+from nautilus_trader.model.events cimport OrderEvent
+from nautilus_trader.model.events cimport OrderFillEvent
+from nautilus_trader.model.events cimport PositionEvent
+from nautilus_trader.model.identifiers cimport AccountId
+from nautilus_trader.model.identifiers cimport StrategyId
+from nautilus_trader.model.identifiers cimport TraderId
+from nautilus_trader.model.position cimport Position
+from nautilus_trader.trading.strategy cimport TradingStrategy
+
+
+cdef class ExecutionEngine:
+    cdef Clock _clock
+    cdef UUIDFactory _uuid_factory
+    cdef LoggerAdapter _log
+    cdef ExecutionClient _exec_client
+    cdef dict _registered_strategies
+
+    cdef readonly TraderId trader_id
+    cdef readonly AccountId account_id
+    cdef readonly ExecutionDatabase database
+    cdef readonly Account account
+    cdef readonly Portfolio portfolio
+    cdef readonly int command_count
+    cdef readonly int event_count
+
+    # -- COMMANDS --------------------------------------------------------------------------------------
+    cpdef void register_client(self, ExecutionClient exec_client) except *
+    cpdef void register_strategy(self, TradingStrategy strategy) except *
+    cpdef void deregister_strategy(self, TradingStrategy strategy) except *
+    cpdef void execute_command(self, Command command) except *
+    cpdef void handle_event(self, Event event) except *
+    cpdef void check_residuals(self) except *
+    cpdef void reset(self) except *
+
+    # -- QUERIES ---------------------------------------------------------------------------------------
+    cpdef list registered_strategies(self)
+    cpdef bint is_strategy_flat(self, StrategyId strategy_id)
+    cpdef bint is_flat(self)
+
+    # --------------------------------------------------------------------------------------------------
+    cdef void _execute_command(self, Command command) except *
+    cdef void _handle_event(self, Event event) except *
+    cdef void _handle_order_cancel_reject(self, OrderCancelReject event) except *
+    cdef void _handle_order_event(self, OrderEvent event) except *
+    cdef void _handle_order_fill(self, OrderFillEvent event) except *
+    cdef void _handle_position_event(self, PositionEvent event) except *
+    cdef void _handle_account_event(self, AccountStateEvent event) except *
+    cdef void _position_opened(self, Position position, StrategyId strategy_id, OrderEvent event) except *
+    cdef void _position_modified(self, Position position, StrategyId strategy_id, OrderEvent event) except *
+    cdef void _position_closed(self, Position position, StrategyId strategy_id, OrderEvent event) except *
+    cdef void _send_to_strategy(self, Event event, StrategyId strategy_id) except *
+    cdef void _reset(self) except *
