+++ nautilus_trader/live/execution_client.pyx
@@ -14,8 +14,8 @@
 # -------------------------------------------------------------------------------------------------
 
 from nautilus_trader.common.clock cimport Clock
-from nautilus_trader.common.execution cimport ExecutionClient
-from nautilus_trader.common.execution cimport ExecutionEngine
+from nautilus_trader.common.execution_client cimport ExecutionClient
+from nautilus_trader.common.execution_engine cimport ExecutionEngine
 from nautilus_trader.common.logging cimport Logger
 from nautilus_trader.common.uuid cimport UUIDFactory
 from nautilus_trader.core.correctness cimport Condition
