+++ tests/unit_tests/backtest/test_backtest_execution.py
@@ -25,8 +25,8 @@ from nautilus_trader.backtest.logging import TestLogger
 from nautilus_trader.backtest.models import FillModel
 from nautilus_trader.backtest.uuid import TestUUIDFactory
 from nautilus_trader.common.data import DataClient
-from nautilus_trader.common.execution import ExecutionEngine
-from nautilus_trader.common.execution import InMemoryExecutionDatabase
+from nautilus_trader.common.execution_database import InMemoryExecutionDatabase
+from nautilus_trader.common.execution_engine import ExecutionEngine
 from nautilus_trader.common.portfolio import Portfolio
 from nautilus_trader.model.enums import OrderSide
 from nautilus_trader.model.events import OrderFilled
