+++ tests/unit_tests/trading/test_trading_trader.py
@@ -24,8 +24,8 @@ from nautilus_trader.backtest.execution import BacktestExecClient
 from nautilus_trader.backtest.logging import TestLogger
 from nautilus_trader.backtest.models import FillModel
 from nautilus_trader.backtest.uuid import TestUUIDFactory
-from nautilus_trader.common.execution import ExecutionEngine
-from nautilus_trader.common.execution import InMemoryExecutionDatabase
+from nautilus_trader.common.execution_database import InMemoryExecutionDatabase
+from nautilus_trader.common.execution_engine import ExecutionEngine
 from nautilus_trader.common.portfolio import Portfolio
 from nautilus_trader.model.enums import BarAggregation
 from nautilus_trader.model.enums import ComponentState
